tests: Move `missing_block_headers_will_be_filled_in` to blocks
